feat: implement Native Mirror synchronization and unified UserProgressEntity
Unified word and grammar study states into UserProgressEntity. Implemented SupabaseSyncManager with Realtime support for seamless cross-platform syncing. Standardized ISO-8601 date handling and resolved Hilt DI issues.

+++ Nemo/core/domain/src/main/java/com/jian/nemo/core/domain/model/Grammar.kt
@@ -38,8 +38,8 @@ data class Grammar(
     // ========== SRS 复习字段 (FSRS 6) ==========
     override val repetitionCount: Int = 0,
     override val interval: Int = 0,
-    override val stability: Float = 0f,
-    override val difficulty: Float = 0f,
+    override val stability: Double = 0.0,
+    override val difficulty: Double = 0.0,
     override val nextReviewDate: Long = 0,
     override val lastReviewedDate: Long? = null,
     override val firstLearnedDate: Long? = null,
